Add toolbar, fix symbol input, improve trendlines, add close position button
UI Improvements:
- Add toolbar with drawing tools (None/HLine/Trend), color selection, line styles
- Add SMA, EMA, Bollinger Bands indicator controls with period inputs
- Add "Clear Lines" button to remove all drawings for current symbol
- Add close (X) button on open positions to sell entire position
- Support Command key (macOS) as alternative to Ctrl for hotkeys

Bug Fixes:
- Fix symbol input focus - now properly focuses when clicking "Enter Symbol"
- Fix market data not loading - call load_symbol() when symbol entered
- Fix trendlines snapping to candle centers - now use exact click position

Technical Changes:
- Change TrendLine candle_start/end from int to float for precise positioning
- Update database schema to store trendline positions as REAL
- Add m_edit_frames counter for reliable input focus handling

+++ chart_widget.h
@@ -72,7 +72,7 @@ private:
     int m_dragging_trendline;
     int m_dragging_trendline_point;
     bool m_trendline_drawing;
-    int m_trendline_start_candle;
+    float m_trendline_start_candle;  // Float for exact positioning
     float m_trendline_start_price;
 
     // Calculated indicators
